requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.1
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
cloudinary==1.10.0
gunicorn==19.9.0
django-heroku==0.3.1
whitenoise==4.1
django-rest-swagger==2.2.0                     
django-guardian==1.4.9             
Pillow==5.4.1             
djangorestframework-jwt==1.11.0
django-widget-tweaks==1.4.5
